image-service/requirements.txt
fastapi>=0.110.0
uvicorn[standard]>=0.24.0
sqlalchemy>=2.0.23
psycopg2-binary>=2.9.9
pydantic>=2.5.0
redis>=5.0.1
kafka-python>=2.0.2
httpx>=0.25.2
mcp>=1.0.0
pytest>=7.4.3
pytest-asyncio>=0.21.1
pillow>=10.1.0
